# jobnow-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobNow API",
        "description": "API documentation for JobNow - Moroccan Job Portal Platform. This API provides endpoints for job seekers and companies to manage job offers, applications, and user profiles.",
        "contact": {
            "name": "JobNow Support",
            "email": "support@jobnow.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.jobnow.ma",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/candidat/job-alerts": {
            "get": {
                "tags": [
                    "Job Alerts"
                ],
                "summary": "Get all job alerts for authenticated candidat",
                "description": "Returns list of job alerts configured by the candidat",
                "operationId": "7850e085d69bf207c77f857d90b74558",
                "responses": {
                    "200": {
                        "description": "Job alerts retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Job Alerts"
                ],
                "summary": "Create a new job alert",
                "description": "Create a job alert with specific criteria",
                "operationId": "5804075fef55e793328c5626d9b7dd87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "keywords": {
                                        "type": "string",
                                        "example": "developer"
                                    },
                                    "contract_types": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "locations": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "daily",
                                        "enum": [
                                            "instant",
                                            "daily",
                                            "weekly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job alert created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidat/job-alerts/{id}": {
            "put": {
                "tags": [
                    "Job Alerts"
                ],
                "summary": "Update a job alert",
                "description": "Update an existing job alert",
                "operationId": "4f569df2b2a12969add1d6947806a687",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job alert updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Job Alerts"
                ],
                "summary": "Delete a job alert",
                "description": "Delete an existing job alert",
                "operationId": "d81ed94a84028beec518b70fc998650f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job alert deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}